refactor: rename useHeaderMenu file to useSidebarMenu
